""Optimización""

+++ Proyecto/src/main/java/org/example/Main.java
@@ -118,8 +118,8 @@ public class Main {
 
         in = new Scanner(System.in);
 
-        long numCasos = in.nextLong();
-        for (long i = 0; i < numCasos; i++)
+        int numCasos = in.nextInt();
+        for (int i = 0; i < numCasos; i++)
             casoDePrueba();
     }
 }
